fix: arrumando verify email e cpf

+++ src/main/java/sptech/school/v2/cleanarch/core/application/gateways/student/StudentQueryGateway.java
@@ -17,7 +17,7 @@ public interface StudentQueryGateway {
 
     Optional<Student> findByEmail(String email);
 
-    Optional<Long> findIdByEmail(String email);
+    Optional<Integer> findIdByEmail(String email);
 
-    Optional<Long> findByCpf(String cpf);
+    Optional<Integer> findByCpf(String cpf);
 }
