+++ src/main/java/sptech/school/v2/cleanarch/core/application/gateways/teacher/TeacherQueryGateway.java
@@ -22,7 +22,7 @@ public interface TeacherQueryGateway {
 
     Optional<Teacher> findByEmail(String email);
 
-    Optional<Long> findIdByEmail(String email);
+    Optional<Integer> findIdByEmail(String email);
 
-    Optional<Long> findByCpf(String cpf);
+    Optional<Integer> findByCpf(String cpf);
 }
